Babbys first shader

+++ mandelbrot/src/range.rs
@@ -1,17 +1,18 @@
 #[derive(Clone, Copy)]
+#[repr(C, align(16))]
 pub struct Range {
-    min: f64,
-    max: f64,
+    min: f32,
+    max: f32,
 }
 
 impl Range {
-    pub fn new(min: f64, max: f64) -> Self {
+    pub fn new(min: f32, max: f32) -> Self {
         Self { min, max }
     }
 
-    pub fn scale(input: &Self, value: f64, output: &Self) -> f64 {
-        let input_size = f64::abs(input.max - input.min);
-        let output_size = f64::abs(output.max - output.min);
+    pub fn scale(input: &Self, value: f32, output: &Self) -> f32 {
+        let input_size = f32::abs(input.max - input.min);
+        let output_size = f32::abs(output.max - output.min);
         (input.max * output.min - input.min * output.max + value * output_size) / input_size
     }
 }
