+++ mandelbrot/src/vector.rs
@@ -1,11 +1,11 @@
 #[derive(Debug, Clone)]
 pub struct Vector {
-    pub x: f64,
-    pub y: f64,
+    pub x: f32,
+    pub y: f32,
 }
 
 impl Vector {
-    pub fn new(x: f64, y: f64) -> Self {
+    pub fn new(x: f32, y: f32) -> Self {
         Self { x, y }
     }
 }
